Stop using looping templates as they don't scale to an arbitrary dimension configuration
