+++ src/binary/binary_op.cc
@@ -36,14 +36,17 @@ struct BinaryOpImplBody<VariantKind::CPU, OP_CODE, CODE, DIM> {
                   const Rect<DIM>& rect,
                   bool dense) const
   {
+    const size_t volume = rect.volume();
     if (dense) {
-      size_t volume = rect.volume();
-      auto outptr   = out.ptr(rect);
-      auto in1ptr   = in1.ptr(rect);
-      auto in2ptr   = in2.ptr(rect);
+      auto outptr = out.ptr(rect);
+      auto in1ptr = in1.ptr(rect);
+      auto in2ptr = in2.ptr(rect);
       for (size_t idx = 0; idx < volume; ++idx) outptr[idx] = func(in1ptr[idx], in2ptr[idx]);
     } else {
-      CPULoop<DIM>::binary_loop(func, out, in1, in2, rect);
+      for (size_t idx = 0; idx < volume; ++idx) {
+        auto p = pitches.unflatten(idx, rect.lo);
+        out[p] = func(in1[p], in2[p]);
+      }
     }
   }
 };
