+++ src/binary/binary_red_template.inl
@@ -15,7 +15,7 @@
  */
 
 #include "binary/binary_op_util.h"
-#include "point_task.h"
+#include "pitches.h"
 
 namespace legate {
 namespace numpy {
