+++ src/nullary/fill_template.inl
@@ -15,7 +15,7 @@
  */
 
 #include "arg.h"
-#include "point_task.h"
+#include "pitches.h"
 
 namespace legate {
 namespace numpy {
