+++ src/unary/scalar_unary_red_template.inl
@@ -15,7 +15,7 @@
  */
 
 #include "unary/unary_red_util.h"
-#include "point_task.h"
+#include "pitches.h"
 
 namespace legate {
 namespace numpy {
