+++ src/unary/unary_red_template.inl
@@ -16,7 +16,7 @@
 
 #include "unary/unary_red_util.h"
 #include "arg.h"
-#include "point_task.h"
+#include "pitches.h"
 
 namespace legate {
 namespace numpy {
